UI Testing - DONT USE THIS PROJECT
Added another project where ive started to update the UI. This push has the base in, all calls and cooldowns needa be redone. THIS PROJECT DOESNT WORK. USE THE OLD PROJECT TO SHOW LECTURES. ILL WORK ON THIS MORE LATER ON. This is very much for me to save it plus as a reminder for when i fully implement it on how to add certain things. Merging was a pain xD