Added room array types (randomization is still broken tho)
In the get random room, getting a random room sometimes returns NONE. Maybe because the array size is 0 (1 element). For now I am hard sending the first element (since there is only one). But this might break if there are going to be multiple rooms and it still sends none. I HATE UE4